@@ -100,16 +49,12 @@ setup(
     keywords="flask plugin",
     description="Load and run plugins for your Flask application",
     long_description=_get_readme(),
-    test_suite="setup.test_suite",
-    tests_require=["Flask>=0.11", "Flask-Classful>=0.14.0"],
     packages=["flask_pluginkit"],
     include_package_data=True,
     zip_safe=False,
-    python_requires=">=2.7,!=3.0.*,!=3.1.*,!=3.2.*,!=3.3.*,!=3.3.*,!=3.5.*",
-    install_requires=["Flask>=0.12", "semver>=2.10.0"],
-    cmdclass={
-        "publish": PublishCommand,
-    },
+    python_requires=">=3.8",
+    install_requires=["Flask>=3.0.0", "semver>=3.0.0"],
+    extras_require={"redis": ["redis>=5.0.0"]},
     classifiers=[
         "Development Status :: 4 - Beta",
         "Framework :: Flask",
