@@ -9,48 +9,18 @@
     :license: BSD 3-Clause, see LICENSE for more details.
 """
 
-from sys import version_info
+import urllib.request as urllib2
+from urllib.parse import urlsplit, parse_qs
 
-PY2 = version_info[0] == 2
 
-if PY2:  # pragma: nocover
+def iteritems(d: dict):
+    return iter(d.items())
 
-    import urllib2
-    from urlparse import urlsplit, parse_qs
 
-    def iteritems(d):
-        return d.iteritems()
+def itervalues(d: dict):
+    return iter(d.values())
 
-    def itervalues(d):
-        return d.itervalues()
 
-    text_type = unicode
-    string_types = (str, unicode)
-    integer_types = (int, long)
-
-else:  # pragma: nocover
-
-    import urllib.request as urllib2
-    from urllib.parse import urlsplit, parse_qs
-
-    def iteritems(d):
-        return iter(d.items())
-
-    def itervalues(d):
-        return iter(d.values())
-
-    text_type = str
-    string_types = (str,)
-    integer_types = (int,)
-
-__all__ = [
-    "PY2",
-    "urllib2",
-    "urlsplit",
-    "parse_qs",
-    "iteritems",
-    "itervalues",
-    "text_type",
-    "string_types",
-    "integer_types",
-]
+text_type = str
+string_types = (str,)
+integer_types = (int,)
